Removed classification output in the absence of any output spikes. Fixed memory leak in batch-loop when using aedat input. Fixed keras compatibility issue when loading model trained on different platform.

+++ snntoolbox/datasets/aedat/DVSIterator.py
@@ -101,7 +101,7 @@ class DVSIterator(object):
             # Get class label, which is the same for all events in a sequence.
             self.y_b = np.broadcast_to(to_categorical(
                 [self.labels[self.dvs_sample_idx]], self.num_classes),
-                (self.batch_size, self.num_classes))
+                (self.batch_size, self.num_classes)).astype('float32')
 
             # Generate frames from events.
             self.frames_from_sequence = get_frames_from_sequence(
